move data related to datable
out of byte stream
