+++ singularity_sap/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod byte_stream;
+pub mod datable;
 pub mod packet;
 pub mod universal_stream;
